pyproject.toml
[tool.poetry]
name = "trade"
version = "0.1.0"
description = ""
authors = ["Pavel14701 <pavelkutia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pydantic = "^2.10.4"
pytelegrambotapi = "^4.25.0"
python-okx = "^0.3.5"
redis = "^5.2.1"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
pandas-ta = "^0.3.14b0"
matplotlib = "^3.10.0"
pandas-flavor = "^0.6.0"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
